fix(mdoc): bind predicate claim names when generating show inputs
generate-mdoc-inputs.ts called generateShowInputs with seven arguments,
omitting the eighth. Its own comment says what that costs: "Default 0
leaves the binding vacuous for flows with no attribute names (JWT); mdoc
callers pass real hashes." It also never set predicateClaimNames or
predicateClaimNameLens.

So claimIdentifierHashes came out ["0","0"] and the Show circuit refused
the witness at eval-predicates.circom:129, where an active predicate must
name a non-zero identity or it would wildcard-match a padding slot.
Witness generation printed "Failed assert in EvalPredicates" and carried
on, so the failure only surfaced later as InvalidSumcheckProof out of
verify. `mdoc benchmark` now reports expressionResult: true.

Both slots get identities, not just the one the predicate names:
MdocCircuit::shared() commits the mdoc circuit's own claimIdentifierHashes,
so Show has to carry the same values or comm_W_shared will not match.

hashClaimName has to run the HashBytesToFieldWithLen circuit — the hash is
over secq256r1 and no JS Poseidon matches — so it and witnessIndices moved
from tests/common into src, where the generators can reach them.
tests/common re-exports both.
