+++ wallet-unit-poc/circom/inputs/mdoc/default.json
@@ -98,143 +98,143 @@
     "0",
     "88",
     "32",
+    "150",
+    "44",
+    "224",
+    "7",
+    "25",
+    "188",
+    "97",
     "136",
-    "45",
-    "37",
-    "213",
-    "108",
-    "175",
-    "108",
-    "129",
-    "186",
-    "94",
-    "193",
-    "162",
-    "33",
-    "175",
-    "101",
-    "155",
-    "158",
-    "240",
-    "60",
-    "229",
-    "98",
-    "180",
-    "193",
-    "248",
-    "184",
-    "104",
+    "200",
+    "208",
+    "17",
+    "42",
+    "36",
+    "39",
+    "22",
+    "53",
+    "100",
+    "53",
+    "226",
+    "249",
+    "0",
+    "234",
+    "177",
+    "200",
+    "208",
+    "135",
+    "123",
     "80",
-    "164",
-    "132",
-    "190",
-    "167",
-    "162",
+    "137",
+    "3",
+    "83",
+    "119",
     "1",
     "88",
     "32",
-    "245",
-    "217",
-    "174",
-    "136",
-    "221",
-    "221",
-    "128",
-    "208",
-    "241",
-    "188",
-    "175",
-    "4",
-    "107",
-    "98",
-    "208",
-    "239",
-    "245",
-    "212",
-    "172",
-    "101",
-    "170",
-    "110",
-    "221",
-    "131",
-    "160",
-    "147",
+    "16",
+    "52",
+    "100",
+    "7",
+    "92",
+    "86",
+    "114",
+    "140",
+    "161",
     "28",
-    "232",
-    "204",
-    "41",
-    "137",
-    "219",
+    "16",
+    "24",
+    "50",
+    "74",
+    "170",
+    "51",
+    "13",
+    "80",
+    "168",
+    "227",
+    "222",
+    "169",
+    "220",
+    "197",
+    "162",
+    "20",
+    "113",
+    "174",
+    "96",
+    "2",
+    "255",
+    "119",
     "2",
     "88",
     "32",
-    "223",
-    "20",
-    "187",
-    "241",
-    "96",
-    "0",
-    "11",
-    "144",
-    "96",
-    "75",
-    "216",
-    "229",
+    "28",
+    "220",
+    "252",
     "29",
-    "149",
-    "247",
-    "16",
-    "201",
-    "114",
-    "124",
-    "99",
-    "73",
-    "2",
-    "21",
-    "161",
-    "95",
-    "212",
-    "36",
-    "54",
-    "137",
-    "126",
+    "12",
     "105",
-    "107",
+    "144",
+    "14",
+    "35",
+    "243",
+    "140",
+    "210",
+    "109",
+    "142",
+    "88",
+    "77",
+    "97",
+    "125",
+    "228",
+    "235",
+    "226",
+    "199",
+    "170",
+    "230",
+    "180",
+    "167",
+    "145",
+    "24",
+    "185",
+    "216",
+    "126",
+    "16",
     "3",
     "88",
     "32",
-    "169",
-    "192",
-    "139",
-    "7",
-    "51",
-    "216",
-    "176",
-    "112",
-    "13",
-    "6",
-    "12",
-    "108",
-    "179",
-    "76",
-    "190",
-    "125",
-    "34",
-    "154",
+    "38",
+    "62",
+    "105",
+    "99",
+    "81",
+    "73",
+    "84",
+    "174",
+    "33",
+    "19",
+    "121",
+    "104",
+    "205",
+    "82",
+    "65",
+    "138",
+    "188",
+    "18",
+    "57",
+    "233",
+    "99",
+    "80",
+    "188",
+    "194",
+    "67",
     "35",
-    "254",
-    "50",
-    "43",
-    "143",
-    "167",
-    "27",
-    "13",
-    "216",
-    "137",
-    "234",
-    "94",
-    "13",
-    "42",
+    "116",
+    "29",
+    "77",
+    "49",
+    "119",
+    "184",
     "109",
     "100",
     "101",
@@ -269,73 +269,73 @@
     "33",
     "88",
     "32",
-    "178",
-    "236",
-    "99",
-    "217",
-    "196",
-    "11",
-    "127",
-    "49",
-    "249",
-    "28",
-    "149",
-    "189",
-    "71",
-    "232",
-    "54",
-    "192",
-    "205",
+    "94",
+    "100",
+    "84",
+    "39",
+    "77",
+    "18",
+    "162",
+    "202",
+    "162",
+    "66",
+    "70",
+    "2",
+    "24",
+    "171",
     "155",
-    "216",
-    "88",
-    "238",
-    "50",
-    "92",
-    "72",
-    "239",
-    "151",
-    "37",
-    "252",
-    "251",
-    "248",
-    "205",
-    "51",
+    "233",
+    "14",
+    "124",
+    "119",
+    "98",
+    "87",
+    "211",
+    "128",
+    "42",
+    "215",
+    "86",
+    "146",
+    "26",
+    "100",
+    "0",
+    "85",
+    "60",
     "34",
     "88",
     "32",
-    "73",
-    "200",
-    "158",
-    "9",
-    "17",
-    "6",
-    "190",
-    "104",
-    "14",
+    "144",
+    "102",
+    "57",
+    "183",
+    "144",
+    "41",
+    "196",
+    "207",
+    "247",
+    "111",
+    "188",
     "180",
-    "75",
-    "147",
+    "83",
+    "127",
+    "239",
+    "228",
     "101",
-    "141",
-    "140",
-    "52",
-    "204",
-    "225",
-    "177",
-    "106",
-    "98",
-    "244",
-    "221",
-    "223",
-    "113",
-    "205",
-    "166",
-    "168",
-    "120",
-    "203",
-    "30",
-    "224",
+    "107",
+    "63",
+    "173",
+    "115",
+    "8",
+    "24",
+    "161",
+    "181",
+    "6",
+    "247",
+    "165",
+    "28",
+    "2",
+    "191",
+    "174",
     "32",
     "1",
     "103",
@@ -1794,10 +1794,10 @@
     "0"
   ],
   "messageLength": "512",
-  "pubKeyX": "45780096279540099580372169523634492733210277723006882672599752594212940613105",
-  "pubKeyY": "1873982841752599206672419404089310177920426705132930572845475086828330213897",
-  "sig_r": "53663807971439672893082760105645836153589186790862540676424487130267790436633",
-  "sig_s_inverse": "113454413506268154589594529686884006190897502429854101742245859541380801904863",
+  "pubKeyX": "63437918232238855306497583291635288818111105930655201455951878501026636708455",
+  "pubKeyY": "21325415521430901320833116907488620239060117604851105438616905185609299793489",
+  "sig_r": "5863863533427603672773548011412907738502760435795382367944268382742198590618",
+  "sig_s_inverse": "50387040521617971457974556639318193795036398855298925902922173097080025461357",
   "validUntilPrefixPos": "443",
   "deviceKeyPos": "266",
   "preimages": [
@@ -1882,38 +1882,38 @@
       "109",
       "88",
       "32",
-      "220",
-      "36",
-      "160",
-      "23",
-      "9",
-      "143",
-      "169",
-      "209",
-      "113",
-      "220",
-      "36",
-      "135",
-      "157",
-      "153",
-      "15",
+      "208",
+      "250",
       "13",
-      "64",
-      "72",
-      "112",
-      "199",
-      "113",
-      "96",
-      "241",
-      "78",
-      "197",
-      "36",
-      "206",
-      "4",
-      "190",
-      "184",
-      "253",
+      "62",
+      "92",
+      "23",
+      "8",
+      "147",
+      "162",
+      "18",
+      "54",
       "10",
+      "24",
+      "183",
+      "18",
+      "82",
+      "192",
+      "41",
+      "144",
+      "144",
+      "66",
+      "137",
+      "107",
+      "173",
+      "102",
+      "104",
+      "150",
+      "209",
+      "107",
+      "10",
+      "41",
+      "57",
       "128",
       "0",
       "0",
@@ -2133,38 +2133,38 @@
       "109",
       "88",
       "32",
-      "156",
-      "184",
-      "221",
-      "36",
-      "250",
-      "104",
-      "226",
-      "101",
-      "164",
-      "121",
-      "40",
-      "240",
-      "30",
-      "225",
-      "18",
-      "159",
-      "47",
-      "89",
-      "249",
-      "159",
-      "230",
-      "69",
-      "78",
-      "113",
-      "125",
-      "204",
-      "232",
-      "105",
-      "21",
-      "101",
-      "249",
+      "123",
+      "140",
+      "72",
+      "124",
+      "134",
+      "9",
+      "163",
+      "77",
+      "192",
+      "27",
+      "135",
+      "51",
+      "167",
+      "180",
+      "35",
+      "220",
+      "124",
+      "190",
+      "142",
+      "100",
+      "186",
       "252",
+      "156",
+      "198",
+      "24",
+      "56",
+      "213",
+      "126",
+      "222",
+      "121",
+      "21",
+      "208",
       "128",
       "0",
       "0",
