+++ wallet-unit-poc/circom/inputs/show/mdoc.json
@@ -1,9 +1,9 @@
 {
-  "deviceKeyX": "80929352098951813569124798168880487066121420337930933651491159126377835842867",
-  "deviceKeyY": "33373298079889519317212487560997254297784365529119078120286450859879583915744",
-  "sig_r": "109025008123220026594679300878796554575936391214616830551287458911736301487764",
-  "sig_s_inverse": "3966545963657324204233975211453766716760681501759484500714057463625355272848",
-  "messageHash": "37862498356388042912148829354472141915630435551766785189932101988618263592269",
+  "deviceKeyX": "42694673279552642191187042806125083010497228152227910033759698268161059083580",
+  "deviceKeyY": "65313666944991914965347777519194390709889576669768020359418135426644329021358",
+  "sig_r": "79613982816350629288453710144814404759086950168809349595600613870441974319765",
+  "sig_s_inverse": "10366933025202498831830061317608389299066028857609771534314905618353164317031",
+  "messageHash": "15017424200657936283998099771210413655505041914535745584384853001372452124916",
   "predicateLen": "1",
   "claimValues": [
     "19900715",
@@ -26,10 +26,150 @@
     "19900715"
   ],
   "claimIdentifierHashes": [
-    "0",
+    "41364283474712499967556998153066206194035137330695026994181060321779523354074",
+    "16962448181397288331646941789033872754793974439788940926653115495827119384374"
+  ],
+  "predicateClaimNames": [
+    [
+      "98",
+      "105",
+      "114",
+      "116",
+      "104",
+      "95",
+      "100",
+      "97",
+      "116",
+      "101",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0"
+    ]
+  ],
+  "predicateClaimNameLens": [
+    "10",
     "0"
   ],
-  "predicateClaimIdentifiers": [
+  "predicateRhsClaimNames": [
+    [
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0",
+      "0"
+    ]
+  ],
+  "predicateRhsClaimNameLens": [
     "0",
     "0"
   ],
